LayoutTests/media/modern-media-controls/tracks-button/tracks-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/tracks-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>TracksButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const tracksButton = new TracksButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("tracksButton.element.localName", "button");
shouldBeTrue('tracksButton.element.classList.contains("icon")');
shouldBeTrue('tracksButton.element.classList.contains("tracks")');
shouldBeEqualToString("tracksButton.iconName", `${Icons.Tracks}`);

const fullscreenTracksButton = new TracksButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });

scheduler.frameDidFire = function()
{
    if ([tracksButton, fullscreenTracksButton].some(button => button.width == 0))
        return;

    shouldBeTrue('tracksButton.element.style.webkitMaskImage.includes("macOS/media-selection@")');
    shouldBeTrue('fullscreenTracksButton.element.style.webkitMaskImage.includes("macOS/media-selection-fullscreen@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
